refactor(frontend): organize settings organization hooks
