+++ desktop/src/components/settings/panes/OrganizationPane.tsx
@@ -20,10 +20,10 @@ import {
   OrganizationLogo,
   OrganizationSection,
 } from "@/components/settings/panes/organization/OrganizationLogo";
-import { useActiveOrganization } from "@/hooks/organizations/use-active-organization";
-import { useOrganizationActions } from "@/hooks/organizations/use-organization-actions";
-import { useOrganizationInvitations } from "@/hooks/organizations/use-organization-invitations";
-import { useOrganizationMembers } from "@/hooks/organizations/use-organization-members";
+import { useOrganizationActions } from "@/hooks/access/cloud/organizations/use-organization-actions";
+import { useOrganizationInvitations } from "@/hooks/access/cloud/organizations/use-organization-invitations";
+import { useOrganizationMembers } from "@/hooks/access/cloud/organizations/use-organization-members";
+import { useActiveOrganization } from "@/hooks/organizations/facade/use-active-organization";
 import { buildSettingsHref } from "@/lib/domain/settings/navigation";
 import { useAuthStore } from "@/stores/auth/auth-store";
