+++ desktop/src/components/settings/sidebar/SettingsSidebar.test.tsx
@@ -24,7 +24,7 @@ vi.mock("@/components/support/SupportDialog", () => ({
   },
 }));
 
-vi.mock("@/hooks/settings/use-app-version", () => ({
+vi.mock("@/hooks/access/tauri/app/use-app-version", () => ({
   useAppVersion: () => ({ data: "0.0.0" }),
 }));
