+++ desktop/src/hooks/access/tauri/app/query-keys.ts
@@ -1,0 +1,3 @@
+export function appVersionKey() {
+  return ["settings", "app-version"] as const;
+}
