+++ desktop/src/hooks/access/tauri/shell/query-keys.ts
@@ -1,7 +1,3 @@
-export function appVersionKey() {
-  return ["settings", "app-version"] as const;
-}
-
 export function availableEditorsKey() {
   return ["settings", "available-editors"] as const;
 }
